fixed invest api

+++ property/serializers/property_investment_serializers.py
@@ -3,4 +3,4 @@ from rest_framework import serializers
 
 class PropertyInvestmentSerializer(serializers.Serializer):
     property_id = serializers.UUIDField()
-    amount = serializers.IntegerField()
+    amount = serializers.FloatField()
